+++ lib/oneview-sdk/resource/api200/volume_snapshot.rb
@@ -29,13 +29,13 @@ module OneviewSDK
 
       # Method is not available
       # @raise [OneviewSDK::MethodUnavailable] method is not available
-      def create
+      def create(*)
         unavailable_method
       end
 
       # Method is not available
       # @raise [OneviewSDK::MethodUnavailable] method is not available
-      def update
+      def update(*)
         unavailable_method
       end