+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Suggested release: v4.0.0
 
+#### Bug fixes
+- Fixed issue #132 get_default_settings in API200 is currently an instance method, but it should be a class method
+
 #### Breaking changes:
 - Fixed issue #93 Logical Switch refresh conflict